Move filodb.core.columnstore -> filodb.core.store
