+++ coordinator/src/main/scala/filodb.coordinator/CoordinatorSetup.scala
@@ -3,7 +3,7 @@ package filodb.coordinator
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
 
-import filodb.core.columnstore.ColumnStore
+import filodb.core.store.ColumnStore
 import filodb.core.FutureUtils
 import filodb.core.metadata.MetaStore
 import filodb.core.reprojector._
