+++ core/src/main/scala/filodb.core/store/Analyzer.scala
@@ -1,4 +1,4 @@
-package filodb.core.columnstore
+package filodb.core.store
 
 import scala.concurrent.{Await, ExecutionContext, Future}
 import scala.concurrent.duration._
